+++ glutin-winit/CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+- **Breaking:** Fixed a typo in a type name (`ApiPrefence` -> `ApiPreference`).
+
 # Version 0.3.0
 
 - **Breaking:** Update _winit_ to `0.28`. See [winit's CHANGELOG](https://github.com/rust-windowing/winit/releases/tag/v0.28.0) for more info.